requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
django-crispy-forms==1.9.2
gunicorn==20.0.4
idna==2.10
Pillow==7.2.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
twilio==6.45.3
urllib3==1.25.10
